Split ShipItInvocation into its own class
